WIP: debugging GMMEM

+++ rgbd_tools/utils/Gui.h
@@ -68,7 +68,7 @@ namespace rgbd {
 
         void drawCone(double _initPoint[3], double _endPoint[3], double _aperture, std::string _name, unsigned _viewportIndex = 0);
 
-		void drawEllipse(const double _semiAxis[3], const Eigen::Affine3d &_transformation, std::string _tag, unsigned _n = 10, double r = 1, double g = 1, double b = 1, unsigned _pointSize = 1, unsigned _viewport = 0);
+        void drawEllipse(const float _semiAxis[3], const Eigen::Affine3f &_transformation, std::string _tag, unsigned _n = 10, double r = 1, double g = 1, double b = 1, unsigned _pointSize = 1, unsigned _viewport = 0);
 
         void drawArrow(const pcl::PointNormal &_point, std::string _tag, double _r = 1, double _g = 1, double _b = 1, unsigned _lineWidth = 1, unsigned _viewport = 0);
